// src/types/constants/constants.ts
export enum ERROR_CODES_AUTH {
    REQUIRE_EMAIL = "Email es obligatorio",
    REQUIRE_PASSWORD = "Password es obligatorio",
    ACCOUNT_NO_EXIST = "Cuenta no encontrada",
    REQUIRE_NAME = "Nombre es obligatorio",
    ACCOUNT_EXIST= "Correo ya existe en la base de datos"
}
export enum CODES_AUTH_MESSAGES {
    LOGOUT_SUCCESS = "Desconexión exitosa",
}
export enum ERROR_CODES_RESTAURANT {
    NO_RESTAURANTS = "Error al encontrar restaunrantes",
}